+++ src/parsers/semgrep.py
@@ -7,7 +7,7 @@ import re
 import csv
 import json
 from .parser_tools import idgenerator, parser_writer
-from .parser_tools.language_resolver import resolve_lang
+from .parser_tools.language_resolver import resolve_lang_from_ext
 from .parser_tools.progressbar import SPACE,progress_bar
 from .parser_tools.toolbox import Fieldnames
 
@@ -88,7 +88,7 @@ def parse(fpath, scanner, substr, prepend):
             path = os.path.join(prepend, path).replace('\\', '/')
             
             # Resolve language of the file
-            lang = resolve_lang(os.path.splitext(path)[1])
+            lang = resolve_lang_from_ext(os.path.splitext(path)[1])
             
             # Get CWE
             cwe = finding.get('cwe', [])