+++ changelog.md
@@ -1,7 +1,9 @@
 # Change Log
 
-## [Version 2.8.0](https://github.com/DevinPatel72/Static-Analysis-Results-Parser/releases/tag/v2.8.0) (2026-07-02)
+## [Version 2.8.0](https://github.com/DevinPatel72/Static-Analysis-Results-Parser/releases/tag/v2.8.0) (2026-07-06)
+- Added Flawfinder to the list of supported scanners
 - Added an updater executable that checks for and installs new SARP releases
+- Fixed bug where if a lock is on the outfile in SARP CLI, the waiting time will result in duplicate output of findings.
 - Fixed background coloring in Rule Builder window
 
 ## [Version 2.7.0](https://github.com/DevinPatel72/Static-Analysis-Results-Parser/releases/tag/v2.7.0) (2026-06-30)